First

+++ src/Lapis.QRCode.Imaging.Drawing/GraphicsTextDrawer.cs
@@ -131,7 +131,7 @@ namespace Lapis.QRCode.Imaging.Drawing
 							//l = 1 - (1-l)*10/(-1*tripMatrix[r, c]*Math.Log((1-l)+1.5)/Math.Log(2));
 							//l = 1 - (1-l)/2;
 							//var luax = state.DoString("return 3 + 4")[0];
-							var res = (int)scriptFunc.Call (3, 5).First ();
+							var res = scriptFunc.Call (3, 5).First ();
 							if (tripMatrix[r, c] < -25){
 								if (l < .6){
 									l = .6;
